Expose inference helpers at package level
