+++ hanzitransfer/data/stroke_dataset.py
