+++ hanzitransfer/inference/cvae_inference.py
@@ -3,6 +3,8 @@ import numpy as np
 import tensorflow as tf
 from PIL import Image
 
+from . import load_cvae_decoder
+
 LATENT_DIM = 16
 
 def main():
@@ -12,7 +14,7 @@ def main():
     parser.add_argument("--output", type=str, default="generated.png", help="Output image file")
     args = parser.parse_args()
 
-    decoder = tf.keras.models.load_model("hanzi_cvae_decoder.keras")
+    decoder = load_cvae_decoder()
     radical_vec = tf.keras.utils.to_categorical(args.radical, num_classes=args.num_radicals)
     z = np.random.normal(size=(1, LATENT_DIM))
     decoder_input = np.concatenate([z, radical_vec.reshape(1, -1)], axis=1)
