+++ pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "hanzitransfer"
+version = "0.1.0"
+description = "HanZi style transfer utilities"
+readme = "README.md"
+requires-python = ">=3.8"
+authors = [{name = "Unknown"}]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["hanzitransfer"]
+
+[project.scripts]
+hanzi-ui = "hanzitransfer.ui.front_cnnver2:main"
+cvae-infer = "hanzitransfer.inference.cvae_inference:main"
